FRONT-19 > сделал ограничения по ролям для vue-router + загрузка баланса

+++ client/src/components/Tracks/ratingMixin.js
@@ -3,21 +3,26 @@ import { createRating, deleteRating } from '../../http/ratingApi';
 export default {
   computed: {
     isUserMarkRating() {
-      return !!this.track.usersRating.find(v => v.userId === this.user.id);
+      return !!this.track.usersRating.find(v => v.userId === this.user?.id);
     }
   },
   methods: {
     async addOrRemoveRating() {
-      try {
-        if (this.isUserMarkRating) {
-          await deleteRating(this.track.id);
-          this.track.usersRating = this.track.usersRating.filter(v => v.userId !== this.user.id);
-        } else {
-          const newRating = await createRating(this.track.id)
-          this.track.usersRating.push(newRating);
+      if (this.isAuth) {
+        try {
+          if (this.isUserMarkRating) {
+            await deleteRating(this.track.id);
+            this.track.usersRating = this.track.usersRating.filter(v => v.userId !== this.user?.id);
+          } else {
+            const newRating = await createRating(this.track.id)
+            this.track.usersRating.push(newRating);
+          }
+        } catch (e) {
+          alert(e);
         }
-      } catch (e) {
-        alert(e);
+      } else {
+        this.$router.push('/login');
+        return;
       }
     }
   },
